Only ints were being shown in stacked bar charts - now corrected All chart axes now show fractions if the data includes fractions (and the results are less than 10) unless the user inputs a specific number of dps Meta passed to format function in Comparison Line Chart (this was missed previously)

+++ src/components/util/getStackedChartData.ts
@@ -80,9 +80,9 @@ export default function getStackedChartData(
     const met = d[metric?.name || ''];
 
     if (segments.includes(seg)) {
-      resultMap[axis][seg] = parseInt(met);
+      resultMap[axis][seg] = parseFloat(met);
     } else {
-      resultMap[axis]['Other'] = (resultMap[axis]['Other'] || 0) + parseInt(met);
+      resultMap[axis]['Other'] = (resultMap[axis]['Other'] || 0) + parseFloat(met);
     }
   });
